[TRTLLM-13212][refactor] Rename sampler/kernels to sampler/ops
Align the sampler subpackage with the existing codebase convention for
provider-specific compute implementations (cf. _torch/modules/fused_moe/ops),
where 'ops' denotes swappable backend implementations behind an interface,
reserving 'kernels' for actual CUDA/C++ kernel sources. No behavior change;
updates all imports, the lazy __getattr__ submodule guard, the mypy
static-analysis file list, and doc comments.
