+++ tensorrt_llm/_torch/pyexecutor/sampler/__init__.py
@@ -17,7 +17,7 @@
 The upper-level orchestration (``Sampler`` / ``TorchSampler`` / ``TRTLLMSampler``)
 lives in ``sampler.py`` and depends on operation-level APIs in
 ``sampling_utils.py``. Implementation-specific kernel providers (FlashInfer,
-vanilla/PyTorch, TRT-LLM ops) live under ``kernels/`` and are selected
+vanilla/PyTorch, TRT-LLM ops) live under ``ops/`` and are selected
 internally, never exposed as interchangeable backends to callers.
 
 Public symbols from ``sampler.py`` are re-exported here so existing
@@ -32,7 +32,7 @@ import importlib
 
 # Submodules of this package — never forward these to sampler.py (that would
 # recurse, since accessing e.g. `.sampler` before it is bound re-enters here).
-_SUBMODULES = frozenset({"sampler", "sampling_utils", "kernels"})
+_SUBMODULES = frozenset({"sampler", "sampling_utils", "ops"})
 
 
 def __getattr__(name: str):
